Changes to Site
It has been a while since I have worked on the Freedom Train campaign. It has also been a while since I have made changes to this website. Because of this, I had to change a few things. First of all, I linked the Facebook page and the Twitter page to the website, and I also added a new picture to replace the one that was on Facebook. (I am still trying to figure out why it was taken down. Oh well.)