Added YouTube Page and Updated Timeline
I added a YouTube button and linked the YouTube channel to the website. I also updated the timeline to show that the YouTube channel was made.

I also updated the sitemap since I did a bit more work on the site today. Hopefully, more things are accomplished this time around.