Updates, Updates, Updates
I added the American Freedom Train Campaign Logo to the main page and the promotional letter page. It kind of shows what the whole campaign is about. This is not a logo for commercial use since I do not have a single add on there. Of course, if I have to cite where I got the train from, then I will.