Update americanFreedomTrain2026Logo.jpg
Flipped the train around lol.

Just kidding. I did flip the train around, but I replaced the original image in the center with an image that could be used for commercial use. That was the real issue. The previous image was not for commercial use. It was a cool image, though.