Removed REMs and Started Building a Form
I just removed several unnecessary REMs from the code of the site and added a few more lines that helped build a form. At some point, I will hook it up with FormSubmit.io, a system that allows individuals to send all of the messages of the site to an account where they can all be viewed.

Because of this, I had to update the sitemap by hand.