Finished Building Form and Message Sent Page
I had to look at the original to build a few parts of the new page, but it works great. The logo does have a few issues with staying in one position, but that is more of a graphics issue, not a website issue. Otherwise, this should work relatively well.