Commenting Out Logo and Improving Message Sent Announcement
I basically removed the logo because I did not know if it was safe for me to have it on there. One of the images was from online, but I do not know if I'm allowed to use it or not. I may have to find a graphic designer to draw me a logo. Hmm.

Anyways, other than doing that to the pages that have the logo, I did add a "www" next to the 4449.com source. The actual address starts with www.

Also, I wrote some JavaScript code for the Message Sent page that randomly chooses an image and displays it on the screen, also showing the citation. By doing this, every individual may have a different image that they see when they send a message.

There are more changes that I have to make, but that is the start of it.